Simplify CryptoTransfer
